+++ Example1-DOTS-DoofusesMustEat/DOOFUSES_STRIDE_COMPUTESHARP/Svelto/com.sebaslab.svelto.common/DataStructures/DualMemorySupport/NativeStrategy.cs
@@ -45,7 +45,7 @@ namespace Svelto.DataStructures.Native
             {
                 IntPtr pointer = _realBuffer.ToNativeArray(out _);
                 pointer = MemoryUtilities.Realloc<T>(pointer, newSize, _nativeAllocator
-                                                   , (uint) newSize > capacity ? (uint) capacity : newSize
+                                                   , newSize > capacity ? (uint) capacity : newSize
                                                    , copyContent);
                 NB<T> b = new NB<T>(pointer, newSize);
                 _realBuffer    = b;
